requirements.txt
# ======================================================
# Enterprise Knowledge Base Expert Collaborative System
# Requirements - Python 3.11 Locked
# ======================================================

# --- Web Framework ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
python-multipart==0.0.9

# --- Config & Utilities ---
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.4
httpx==0.27.0

# --- LangChain Core (DeepSeek via OpenAI-compatible adapter) ---
langchain==0.2.14
langchain-core==0.2.33
langchain-openai==0.1.22
langchain-community==0.2.12

# --- LangGraph (RedisSaver requires the separate checkpoint package) ---
langgraph==0.2.28
langgraph-checkpoint-redis==0.0.6
redis==5.0.4

# --- LlamaIndex Core ---
llama-index-core==0.10.55
llama-index-llms-openai==0.1.27
llama-index-llms-openai-like==0.1.3   # Required for DeepSeek/custom OpenAI-compatible endpoints
llama-index-readers-file==0.1.33
llama-index-embeddings-huggingface==0.2.3

# --- Local Models (HuggingFace - weights downloaded separately) ---
sentence-transformers==3.0.1
transformers==4.41.2
# IMPORTANT: Install CUDA torch manually before pip install -r requirements.txt
# Command (CUDA 12.8 native build):
#   pip install torch==2.10.0 --index-url https://download.pytorch.org/whl/cu128
# Verify: python -c "import torch; print(torch.cuda.is_available())"
torch==2.10.0

# --- Hybrid Search ---
# BM25 sparse retriever
llama-index-retrievers-bm25==0.2.0
rank_bm25==0.2.2

# --- Reranker (Cross-Encoder, lightweight, no CUDA required) ---
llama-index-postprocessor-sbert-rerank==0.1.5

# --- RAGAS Evaluation ---
ragas==0.1.15
datasets==2.20.0
